feat: 예약 시간 삭제 예외 처리
- 특정 시간에 대한 예약이 존재할 때 해당 시간 삭제 시 예외 발생

+++ src/main/java/roomescape/reservationTime/service/ReservationTimeService.java
@@ -2,6 +2,7 @@ package roomescape.reservationTime.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.reservationTime.domain.ReservationTimeRepository;
 import roomescape.reservationTime.dto.ReservationTimeRequest;
@@ -10,9 +11,11 @@ import roomescape.reservationTime.dto.ReservationTimeResponse;
 @Service
 public class ReservationTimeService {
 
+    private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationRepository reservationRepository, final ReservationTimeRepository reservationTimeRepository) {
+        this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
     }
 
@@ -24,13 +27,16 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTimeById(final Long id) {
+        if (reservationRepository.existByReservationTimeId(id)){
+            throw new IllegalArgumentException("삭제할 수 없는 예약 시간입니다.");
+        }
         int count = reservationTimeRepository.deleteById(id);
         validateIsExistsReservationTimeId(count);
     }
 
     private void validateIsExistsReservationTimeId(final int count) {
         if (count == 0) {
-            throw new IllegalArgumentException("[ERROR] 존재하지 않는 예약 시간입니다.");
+            throw new IllegalArgumentException("존재하지 않는 예약 시간입니다.");
         }
     }
